requirements.txt
Flask
requests
python-dotenv
gunicorn
openai-whisper
# torch==2.0.1+cpu
# torchvision==0.15.2+cpu
# torchaudio==2.0.2+cpu
torch
torchvision
torchaudio

